+++ CHANGELOG.rst
@@ -13,6 +13,8 @@ Programmatic changes
 - Refactored the Parameter I/O widgets and added tests
 - Overhauled the ParameterCollectionMixin and removed unused methods.
 - Added a mix-in class to handle an associated file and type checking centrally.
+- Added a ConfigureBinaryDecodingWidget class which handles the decoding
+  options for raw/binary files centrally.
 
 Bugfixes
 --------